Remove PriorityList sample
Priority lists appear to be lists that display the given number of
'priority' items by default and have a link/button that, when clicked,
reveals the remaining items.

This requires JavaScript which would add the need for either a
more-complicated build system or us to serve a already-bundled GOV.UK
JavaScript file.

Both options are overkill for the moment.